extract bank assessment nad bank folder
